+++ CHANGES.md
@@ -5,6 +5,7 @@
 - [Feature] Throw exceptions with formatted messages.
 - [Feature] Allow formatting of exceptions.
 - [Feature] Target size of type MAY be provided using macro LLAMALOG_LOGLINE_SIZE for larger of smaller buffer requirements.
+- [Feature] Flush function.
 - [Bug] Fixed wrong calculation of alignment for log arguments.
 
 ## v1.0.0